Adds seed data for points and keywords.

+++ db/schema/05_points.sql
@@ -10,5 +10,5 @@ CREATE TABLE points (
   picture VARCHAR(255) DEFAULT 'https://cdn1.vectorstock.com/i/1000x1000/48/70/cityline-vector-21554870.jpg',
   lat DOUBLE PRECISION,
   long DOUBLE PRECISION,
-  category SMALLINT
+  category TINYINT
 );
